move useEffect to orderbook, fix code syntax in login file
