+++ packages/augur-ui/src/modules/market-charts/components/order-book/order-book.tsx
@@ -10,6 +10,7 @@ import {
   SELL,
   SCALAR,
   BINARY_CATEGORICAL_FORMAT_OPTIONS,
+  MIN_ORDER_LIFESPAN,
 } from 'modules/common/constants';
 import { CancelTextButton } from 'modules/common/buttons';
 import Styles from 'modules/market-charts/components/order-book/order-book.styles.less';
@@ -19,6 +20,7 @@ import {
 } from 'modules/types';
 import { createBigNumber } from 'utils/create-big-number';
 import { formatShares } from 'utils/format-number';
+import { NUMBER_OF_SECONDS_IN_A_DAY } from 'utils/format-date';
 
 interface OrderBookSideProps {
   orderBook: QuantityOutcomeOrderBook;
@@ -46,6 +48,9 @@ interface OrderBookProps {
   marketType: string;
   showButtons: boolean;
   orderbookLoading: boolean;
+  expirationTime: number;
+  currentTimeInSeconds: number;
+  loadMarketOrderBook: Function;
 }
 
 const OrderBookSide = ({
@@ -191,10 +196,33 @@ const OrderBook = ({
   marketType,
   showButtons,
   orderbookLoading,
+  expirationTime,
+  currentTimeInSeconds,
+  loadMarketOrderBook,
 }: OrderBookProps) => {
   const [hoverState, setHoverState] = useState({ hoveredOrderIndex: null,
     hoveredSide: null });
   const setHovers = (hoveredOrderIndex: number, hoveredSide: string) => setHoverState({ hoveredOrderIndex, hoveredSide });
+
+  useEffect(() => {
+    const expirationMaxSeconds =
+      expirationTime - currentTimeInSeconds - MIN_ORDER_LIFESPAN;
+    if (
+      expirationMaxSeconds > 0 &&
+      expirationMaxSeconds < NUMBER_OF_SECONDS_IN_A_DAY
+    ) {
+      console.log('Timer', expirationMaxSeconds);
+      const timer = setTimeout(() => {
+        console.log('Timer hit');
+        loadMarketOrderBook();
+      }, expirationMaxSeconds * 1000);
+      return () => {
+        console.log('Timer stop');
+        clearTimeout(timer);
+      };
+    }
+    return () => {};
+  }, [expirationTime]);
 
   return (
     <section className={Styles.OrderBook}>
