+++ packages/augur-ui/src/modules/market-charts/containers/order-book.ts
@@ -3,6 +3,7 @@ import { isEmpty } from "utils/is-empty";
 import OrderBook from "modules/market-charts/components/order-book/order-book";
 import { selectMarket } from "modules/markets/selectors/market";
 import { selectCurrentTimestampInSeconds } from "store/select-state";
+import { loadMarketOrderBook } from 'modules/orders/actions/load-market-orderbook';
 import { ASKS, BIDS } from "modules/common/constants";
 import orderAndAssignCumulativeShares from "modules/markets/helpers/order-and-assign-cumulative-shares";
 
@@ -16,7 +17,10 @@ const mapStateToProps = (state, ownProps) => {
       (outcome) => outcome.id === selectedOutcomeId
     );
 
+  const expirationTime = ownProps.initialLiquidity || !!!orderBook ? 0 : orderBook.expirationTime;
+
   return {
+    expirationTime,
     outcomeName: outcome && outcome.description,
     selectedOutcome: outcome,
     currentTimeInSeconds: selectCurrentTimestampInSeconds(state),
@@ -30,4 +34,20 @@ const mapStateToProps = (state, ownProps) => {
   };
 };
 
-export default connect(mapStateToProps)(OrderBook);
+const mapDispatchToProps = (dispatch) => ({
+  loadMarketOrderBook: marketId => dispatch(loadMarketOrderBook(marketId)),
+});
+
+const mergeProps = (sP: any, dP: any) => {
+
+  return {
+    ...sP,
+    loadMarketOrderBook: () => dP.loadMarketOrderBook(sP.marketId),
+  }
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+  mergeProps
+)(OrderBook);
